requirements.txt
# CORE AI (Pinned to 0.2.x to ensure AgentExecutor exists)
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-text-splitters==0.2.4

# RAG & VECTOR DB
chromadb==0.5.5
pypdf==4.3.1
tiktoken==0.7.0

# TOOLS
yfinance==0.2.43
duckduckgo-search==6.2.11b1
numexpr==2.10.1

# INFRASTRUCTURE
python-dotenv==1.0.1
uvicorn==0.30.6
fastapi==0.112.2